+++ src/test/java/de/dm/intellij/liferay/language/xml/LiferayXmlFileReferenceContributorTest.java
@@ -1,11 +1,38 @@
 package de.dm.intellij.liferay.language.xml;
 
 import com.intellij.codeInsight.completion.CompletionType;
+import com.intellij.openapi.module.Module;
+import com.intellij.openapi.roots.ContentEntry;
+import com.intellij.openapi.roots.ModifiableRootModel;
+import com.intellij.openapi.util.Disposer;
+import com.intellij.openapi.vfs.newvfs.impl.VfsRootAccess;
+import com.intellij.testFramework.LightProjectDescriptor;
 import com.intellij.testFramework.fixtures.BasePlatformTestCase;
+import com.intellij.testFramework.fixtures.DefaultLightProjectDescriptor;
+import com.intellij.util.PathUtil;
+import de.dm.intellij.liferay.language.freemarker.LiferayFtlVariableProviderTest;
+import org.jetbrains.annotations.NotNull;
 
+import java.io.File;
+import java.net.URL;
 import java.util.List;
 
 public class LiferayXmlFileReferenceContributorTest extends BasePlatformTestCase {
+
+    private static final LightProjectDescriptor MY_PROJECT_DESCRIPTOR = new DefaultLightProjectDescriptor() {
+        @Override
+        public void configureModule(@NotNull Module module, @NotNull ModifiableRootModel model, @NotNull ContentEntry contentEntry) {
+            URL resource = LiferayFtlVariableProviderTest.class.getResource("/com/liferay/xsd");
+            String resourcePath = PathUtil.toSystemIndependentName(new File(resource.getFile()).getAbsolutePath());
+            VfsRootAccess.allowRootAccess( Disposer.newDisposable(), resourcePath );
+        }
+    };
+
+    @NotNull
+    @Override
+    protected LightProjectDescriptor getProjectDescriptor() {
+        return MY_PROJECT_DESCRIPTOR;
+    }
 
     private static final String TEST_DATA_PATH = "testdata/de/dm/intellij/liferay/language/xml/LiferayXmlFileReferenceContributorTest";
 
@@ -36,5 +63,13 @@ public class LiferayXmlFileReferenceContributorTest extends BasePlatformTestCase
         myFixture.complete(CompletionType.BASIC, 1);
         List<String> strings = myFixture.getLookupElementStrings();
         assertTrue(strings.contains("custom.xml"));
+    }
+
+    public void testPortletDisplayTemplateXmlTagFileReference() {
+        myFixture.configureByFiles("portlet-display-templates.xml", "files/foo.ftl");
+
+        myFixture.complete(CompletionType.BASIC, 1);
+        List<String> strings = myFixture.getLookupElementStrings();
+        assertTrue(strings.contains("files"));
     }
 }